lambdas/upload_youtube/requirements.txt
boto3>=1.34.100
google-api-python-client>=2.127.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.0
